test(stdio): isolate the frame-tracing tests into their own binary
`async_stdio_variants_trace_metadata_without_frame_bodies` failed roughly
one run in seven on unmodified main, with zero captured DEBUG events where
four were expected.

The three tests in `safe_frame_tracing` assert on what a `tracing`
subscriber captured. They install it with `set_default`, which is
thread-local, but the callsite interest cache behind it is process-wide.
Sharing a binary with forty-odd other tests that drive stdio transports
meant the same DEBUG callsites could be evaluated, and their interest
cached, on a thread with no subscriber installed. After that the frame
events never reached the capture and the assertions counted none.

Cargo gives every `tests/*.rs` its own binary, so moving them out buys
process isolation for the cost of three small helpers. Nothing else in the
new process emits a tracing event.

Measured rather than assumed, 30 runs per configuration:

    shared binary                      3/20 failed
    shared + rebuild_interest_cache    1/20 failed
    isolated                           0/30 failed
    isolated + rebuild_interest_cache  0/30 failed

`rebuild_interest_cache` was the first fix tried and looked convincing at
8/8 before a larger sample caught it still failing. It is not in this
change: isolation alone is sufficient, and keeping it would imply a cause
that is not the one being fixed.

Pure movement otherwise. 45 tests before, 42 in `stdio_loop` and 3 in the
new binary after. The full workspace suite ran 10 consecutive times clean.

Closes #1435.
